add initial API and implementation
This project is based heavily off of the work we've done with
[Rb](https://github.com/redding/rb).

This adds the main CLI and sub-command implementations and the
scripts to release, install, and uninstall the application.

The API is very simiilar to Rb except it works off of the installed
Postgres.app versions found in
`/Applications/Postgres.app/Contents/Versions`. All this does is
modify PATH to include the specified version's bin path. It has
features similar to Rb including: status output, listing available
versions, and `.chpg-version` file support. It also supports auto
updating versions on changing directories so you can use a
`.chpg-version` file to lock in the Postgres version for individual
projects.

I found myself wanting this as I want to lock in certain projects
to a specific version of Postgres and control when/if I upgrade.
However, I don't control the Postgres version requirements in other
projects and need to be able to use a different version on those
projects.